+++ libs/set/set.go
@@ -1,4 +1,4 @@
-package core
+package set
 
 import (
 	"fmt"
