+++ libs/set/set_test.go
@@ -1,4 +1,4 @@
-package core
+package set
 
 import (
 	"fmt"
